track "entered" status in genobject
If we haven't run the generator at all before cleaning
it up, we were trying to run it (to do cleanup).  But
since we haven't run before, this calls the function;
it didn't properly guard itself for this.  But since we
haven't run the generator at all yet, we don't have
any refs that we need to cleanup anyway, so just don't
run it.